requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
apache-beam==2.57.0
array-record==0.5.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==2.2.1
clu==0.0.9
comm==0.2.0
contextlib2==21.6.0
contourpy==1.2.0
crcmod==1.7
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.1.1
dm-tree==0.1.8
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
etils==1.9.2
executing==2.0.1
fastavro==1.9.5
fasteners==0.19
flatbuffers==23.5.26
flax==0.8.4
fonttools==4.47.0
fsspec==2023.12.2
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grain==0.1.0
grpcio==1.64.1
gviz-api==1.10.0
h5py==3.10.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.6
immutabledict==4.2.0
importlib-resources==6.1.1
install==1.3.5
ipykernel==6.28.0
ipython==8.19.0
jax==0.4.28
jaxlib==0.4.28+cuda12.cudnn89
jaxtyping==0.2.33
jedi==0.19.1
Js2Py==0.74
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.6.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.4.0
more-itertools==10.3.0
msgpack==1.0.7
namex==0.0.8
nest-asyncio==1.5.8
numpy==1.26.4
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.2.140
oauthlib==3.2.2
objsize==0.7.0
omegaconf==2.3.0
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
optax==0.2.2
optree==0.12.1
orbax==0.1.9
orbax-checkpoint==0.5.23
orjson==3.10.6
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
promise==2.3
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydot==1.4.2
Pygments==2.17.2
pyjsparser==2.7.1
pymongo==4.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.19.0
rsa==4.9
scipy==1.11.4
seaborn==0.13.2
simple_parsing==0.1.5
six==1.16.0
stack-data==0.6.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-profile==2.15.0
tensorflow==2.15.0
tensorflow-datasets==4.9.6
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorstore==0.1.63
termcolor==2.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
zstandard==0.23.0
